backend/requirements.txt
fastapi
google-auth
google-auth-library
psycopg2
pydantic
sqlalchemy
typing
uvicorn